requirement.txt
streamlit
numpy
faiss-cpu
sentence-transformers
pymupdf
ollama
torch